requirements.txt
Flask==2.0.2
google-cloud-ndb==1.11.1